If supported, non-aligned figures will now also use transform to centre themselves.
This is in order to avoid the following hacky approach for 'figure > a':
Restored specific styles for 'figure > a' that made it an inline-block inside the figure, which were needed to have usable hyperlinks in figures.
This is only needed when transform is not supported.

Moved various rules that based on static lengths into the @media tag for wider screens, and replaced them with rules that set widths to 100% for narrow screens.