Make inline lists fully inline instead of flex
Remove extraneous space between <li> by setting word-spacing to `-1em` instead of using flexbox.

Flexbox does not work properly when the inline list starts in the middle of a line and then has a line break, as it causes the inline list to break at its very start, whereas expected behaviour is to break in its middle.